Factor out c_parser/{source,util}.py and c_symbols/local.py.
